// Code/my-react-app/src/v1/components/MobileViewComponents/ProgramsComponent/enums/enums.ts
// common enums
export enum AdminRole {
  MUSALI_ADMIN = "musaliadmin",
  SUB_ADMIN = "subadmin", // MASJID ADMIN
  SUPER_ADMIN = "superadmin",
  ADMIN = "admin", //DATA ENTRY
}

//deatils page enums
export enum RecurrenceType {
  DAILY = "daily",
  WEEKLY = "weekly",
  NONE = "none",
  RANDOM = "random",
}